+++ src/test/java/com/syaru/ae2craftingoptimizer/engine/Ae2CraftingPlanSidecarsTest.java
@@ -84,6 +84,11 @@ class Ae2CraftingPlanSidecarsTest {
         }
 
         @Override
+        public java.math.BigInteger exactBytes() {
+            return java.math.BigInteger.valueOf(Long.MAX_VALUE).add(java.math.BigInteger.ONE);
+        }
+
+        @Override
         public boolean simulation() {
             return false;
         }